+++ lib/your_snackbar.dart
@@ -1,3 +1,4 @@
 library your_snackbar;
 
 export 'src/widget/show_your_snackbar.dart';
+export 'src/widget/your_snackbar_style.dart';