+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1
 
 * TODO: Describe initial release.
+
+## 1.0.0
+* added safe area
+* added bottom direction
+* added more customizable parameters (duration, height, etc)
+* fixed some errors